stop duplicate configurations

+++ scripts/configure.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env bash
 
+if [[ -f "/clustern-node" ]]
+then
+    exit 0
+fi
+
 wget -O /usr/local/bin/pacapt https://github.com/icy/pacapt/raw/ng/pacapt
 chmod 755 /usr/local/bin/pacapt
 ln -sv /usr/local/bin/pacapt /usr/local/bin/pacman || true
